requirements.txt
Django==1.7.4
djangorestframework==3.0.5
django-filter==0.9.2
selenium==2.45.0
gunicorn==19.3.0
ipython==3.0.0
readline==6.2.4.1
requests==2.6.0
django-cors-headers==1.0.0
django-filters==0.1.0
socketIO-client==0.7.2